third_party/WebKit/LayoutTests/transforms/inline-in-transformed-multicol.html

<!DOCTYPE html>
<script src="../resources/ahem.js"></script>
<style>
    body { margin:0; }
</style>
<div style="columns:3; width:100px; height:100px; transform:rotate(90deg); font:16px/1 Ahem; background:yellow;">
    <span id="elm">XXX</span>
</div>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script>
    test(function() {
        var rects = document.getElementById("elm").getClientRects();
        assert_equals(rects.length, 1);
        var rect = rects[0];
        assert_equals(rect.left, 84, "left");
        assert_equals(rect.top, 0, "top");
        assert_equals(rect.right, 100, "right");
        assert_equals(rect.bottom, 48, "bottom");
    }, "getClientRects of inline in transformed multicol");
</script>
